- change the format of a service object to be a list and not an array 	- implemented ISO 8901 for TTL 	- fixed a few bugs

+++ lookup/lookup-service/src/main/java/net/es/lookup/database/ServiceDAOMongoDb.java
@@ -136,7 +136,7 @@ public class ServiceDAOMongoDb {
 		String errormsg;
 		
 		String uri = message.getURI();
-        int ttl = message.getTTL();
+        long ttl = message.getTTL();
 		BasicDBObject query = new BasicDBObject();
 		//TODO: add check to see if only one elem is returned
 		query.put("uri", uri);
